Unset $BAUD when initializing the shell
Zsh checks the BAUD environment variable for a speed of the terminal and
uses this information when redrawing the command line.  The thing is,
most of the time, the value stored in BAUD is _way_ off the actual
baud-rate of the actual terminal, so it winds up postponing redraws that
could have been easily performed without problems.  Unsetting BAUD
altogether is the easiest solution for getting quick redraws of the
command line, so that’s what we do.