ratpoisonrc: Add new keymap to activate windows by class
Although ^m^m is probably the most useful binding that I never use, these
bindings make it possible to select among a set of windows that have a
given class.  This is nice as one can easily select the first XTerm window,
whatever number it has, or the first Firefox window, or whatever.