screenrc: Changed escape and updated to new logging framework
Using ^A as an escape was getting a bit boring to type, as one has to
cringe one’s left hand to use it.  Using ^T, where both hands are used is
much more ergonomic, and it works much better with Emacs-enabled programs.

As metalog doesn’t seem to play well with hald/dbus (using 99% CPU-time)
we’ve switched to using syslog-ng instead.  That means we need to monitor
/var/log/messages instead of the old paths metalog was using.