mplayer/config: Remove srate setting
Previously, srate was being set, as alsa has a hard time figuring out the
rate on its own.  However, setting it for mplayer explicitly instead of in
the alsa configuration-file was a bad choice.  This has now been remedied
so that it works for all applications and the explicit setting for mplayer
can now be removed.