zsh/.zshrc: Add “tv” alias for displaying video on television
I finally managed to get the ATI driver to display video on my television.
It’s a dual-head setup, so one needs to mess about a bit with the $DISPLAY
environment-variable and also choose my secondary sound-card for audio
output.  The “tv” alias does all that, which is a real convenience – as
befits aliases…