zsh/.zshrc: Add self-insert-redir to fix “disappearing” spaces for slashes.
Zsh normally substitutes a trailing slash for a space after a completion,
depending on what character is typed next.  It doesn’t always get it right,
though; and one such case is when one of the following characters are used:
‘|’, ‘<’, ‘>’, and ‘&’.  It doesn’t change the command in any way, but it
looks kind of weird.  So this patch sees to it that a space is added for
these characters as well.  The function self-insert-redir() can of course
be used for other characters as well, not that I know of any that need this
extra handling, besides those that we’ve already dealt with.