vim/vimrc: Add mappings to move faster in completion menu.
This patch maps the number keys to move the number they represent entries
down the completion menu (‘0’ being ten).  This speeds up access in the
menu a lot, but Vim seems to be overzealous in its redrawing routine and
redraws all entries, which is strange, as I have 'lazyredraw' set.  Got to
talk to Bram about that.