zsh/functions/_git: Remove GIT_DIR variable and get it from git instead.
Previously, I wrongly reasoned that a lot of commands would want to know
where .git was.  Also, I did it in a totally wrong fashion, which this
patch remedies.  now the completions that need to look in GIT_DIR properly
ask git for it.

Also included is a fix for __git_heads_or_tags.

Reported by Sergey Vlasov <vsu@altlinux.ru>.