zsh/functions/d: Try to make output a bit cleaner.
This patch tries (but fails) to make the output of “d” a bit cleaner, by
removing common prefixes in the paths listed.  The reasoning is that you
usually delve in and out of a given subtree more than you go from sibling
to sibling, thus the output will often contain prefixes that are common
between a subsequence of the path list.

The problem is that it isn’t always correct, and I haven’t had time to fix
it.