zsh/functions/_git: Update _git-add() file argument handling.
The last patch was wrong.  What git-add cares about is modified files and
other files, that is, untracked files.  Any ignored files of those
varieties should also be included, provided that the “-f” option has been
given to git-add.

Also, the __git_ignored_files() function is useless, as “--ignored” only
produces interesting output when applied together with another specifier,
as it modifies that specifiers normal output to include any ignored files
of that type.  If we want the ignored files as well, simply pass
“--ignored” on as an argument to the __git_*_files() function, as has been
done in this patch.