zsh/functions/_git: Update git-rev-list, git-log, and git-shortlog.
The three commands git-rev-list, git-log, and git-shortlog all take the
same subset of arguments, so factor that out.  I’m not quite sure if this
solution is the one I prefer, but it will do for now.  I’d prefer it if
options specific to each command would be listed separately, and perhaps
it would be nice if options like those for setting up diffs could be under
its own heading.