zsh/functions/_git: Update git-config completion.
This update is quite big, as the whole handling of the names and values to
git-config has been vastly improved.  It’s all context sensitive and
basically does the right thing for every value.  It’s been a lot of work,
so hopefully it’ll pay off.  I’m personally quite proud with how I solved
the handling of color values, for example.