screenrc: Remove bold-to-bright mapping.
Previously we needed to map bold to bright, as many applications still
used that method for representing bright colors.  This has changed (and
the applications I use has also changed) and the screen-256color terminfo
definition also seems to deal with this appropriately, so there’s no
reason to have screen do it internally any more.