Switch to xmonad WM.
xmonad has proven to be just as stable and usable as ratpoison, while
adding some nice features that I simply can’t live without after using it
for a couple of weeks.

Thanks for the wonderful times, ratpoison, but it’s time for both me and
you to move on.