zsh/zshrc: Add “foldl” and “foldr” aliases
The “foldl” alias allows you to execute a command for each of the
following arguments.  This is useful together with commands that normally
only take one argument, like tex(1).

The “foldr” alias allows you to execute a number of commands with one
argument.  This is useful together with init files, though it’s nowhere as
useful as “foldl”.