Add cd-from-list, cd-to-alternate-directory, and up-directory
These are Zle functions that are useful for dealing with the current
directory.

The cd-from-list function displays the directory stack and allows you to
choose a directory to change to.

The cd-to-alternate-directory functions switches to the alternate
(previous) directory.

The up-directory functions changes directory to the parent directory of
the current directory.