Map <C-Z> in normal mode to launching shell
Instead of a no-op, I think having <C-Z> mappeed to launching a shell
can actualy be very useful.  This way we can easily run git, move files,
and so on “straight from” Vim.  I think this is a step in the right
direction if we’re heading towards Emacs in the end.